update: props周りの整備
