+++ components/blogs/form/FormMarkdown.vue
@@ -6,11 +6,17 @@
       :show-warning="showWarning"
       class="mb-3"
     />
+    <form-asset class="mb-3" :on-file-picked="onFilePicked" />
     <mavon-editor
       v-model="description"
       :toolbars="toolbarsOption"
       language="ja"
-      :xss-options="{}"
+      :xss-options="{
+        whiteList: {
+          video: ['width', 'height', 'controls'],
+          source: ['src', 'type']
+        }
+      }"
       default-open="edit"
       :autofocus="false"
       class="markdown-body min-h-300 w-full bg-white"
@@ -19,21 +25,29 @@
 </template>
 
 <script lang="ts">
-import { Component, Vue, VModel, Prop } from 'nuxt-property-decorator'
-import FormLabel from '@/components/works/form/FormLabel.vue'
+import { Component, Vue, Prop, VModel } from 'nuxt-property-decorator'
 import { workPostStore } from '@/store'
+import FormLabel from '@/components/works/form/FormLabel.vue'
+import FormAsset from '@/components/blogs/form/FormAsset.vue'
 
 @Component({
   components: {
-    FormLabel
+    FormLabel,
+    FormAsset
   }
 })
 export default class FormMarkdown extends Vue {
-  @VModel({ type: String })
+  @VModel({
+    type: String,
+    required: true
+  })
   description!: string
 
   @Prop({ type: Boolean, required: true })
   showWarning!: boolean
+
+  @Prop({ type: Function, required: true })
+  onFilePicked!: (e: Event) => void
 
   changeBlockUnloadState() {
     workPostStore.changeIsBlockUnload()
