+++ components/works/WorksForm.vue
@@ -51,7 +51,7 @@
       "
     >
       <form-markdown
-        v-model:description="workData.description"
+        v-model="workData.description"
         :show-warning="showRequiredWarning.descriptionEmpty"
         class="mb-1"
       />
